+++ CHANGELOG.md
@@ -1,21 +1,28 @@
 # Changelog
 
 ## [Unreleased]
-- Update libraries
+
+- Add option to remove Assets/Media (contributed by [@anisval](https://github.com/anisval)
+- Update to latest contentful-manangement release
 
 ## [0.5.0]
+
 - Add env option to specify Contentful environment (contributed by @Ikstar)
 - Update dependencies
 
 ## [0.4.2]
+
 - Return exit code 1 when exception is thrown
 
 ## [0.4.1]
+
 - Improve README
 
 ## [0.4.0]
+
 - Add --content-types parameter to delete content types as well
 
 ## [0.3.0]
+
 - Prompt for confirmation before deleting all entries
 - New option `--yes` to override above confirmation